requirements.txt
openai
python-dotenv
pydub 
instabot 
Pillow